backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
scikit-learn
google-generativeai
python-dotenv
